+++ tests/Tubeshade.Server.Tests.Integration/Tasks/TasksTests.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using FluentAssertions.Execution;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging.Abstractions;
 using NodaTime;
 using Npgsql;
 using NUnit.Framework;
@@ -41,7 +42,6 @@ public sealed class TasksTests(ServerFixture fixture) : ServerTests(fixture)
     }
 
     [Test]
-    [Order(1)]
     public async Task IndexVideoWhileDownloading()
     {
         await using var scope = Fixture.Services.CreateAsyncScope();
@@ -51,54 +51,22 @@ public sealed class TasksTests(ServerFixture fixture) : ServerTests(fixture)
         var channelRepository = scope.ServiceProvider.GetRequiredService<ChannelRepository>();
         var videoRepository = scope.ServiceProvider.GetRequiredService<VideoRepository>();
 
-        Guid channelId;
+        var url = $"https://example.org/{Guid.NewGuid():N}";
         Guid videoId;
 
         await using (var transaction = await connection.OpenAndBeginTransaction())
         {
-            channelId = (await channelRepository.AddAsync(
-                new()
-                {
-                    CreatedByUserId = _userId,
-                    ModifiedByUserId = _userId,
-                    OwnerId = _userId,
-                    Name = Guid.NewGuid().ToString("N"),
-                    StoragePath = string.Empty,
-                    ExternalId = string.Empty,
-                    ExternalUrl = string.Empty,
-                    Availability = ExternalAvailability.Public,
-                },
-                transaction))!.Value;
-
-            videoId = (await videoRepository.AddAsync(
-                new()
-                {
-                    CreatedByUserId = _userId,
-                    ModifiedByUserId = _userId,
-                    OwnerId = _userId,
-                    Name = Guid.NewGuid().ToString("N"),
-                    Description = string.Empty,
-                    Categories = [],
-                    Tags = [],
-                    Type = VideoType.Video,
-                    ChannelId = channelId,
-                    StoragePath = string.Empty,
-                    ExternalId = string.Empty,
-                    ExternalUrl = "https://example.org",
-                    PublishedAt = SystemClock.Instance.GetCurrentInstant(),
-                    RefreshedAt = SystemClock.Instance.GetCurrentInstant(),
-                    Availability = ExternalAvailability.Public,
-                    Duration = Period.FromMinutes(10),
-                },
-                transaction))!.Value;
+            var channelId = await CreateChannel(channelRepository, transaction);
+            videoId = await CreateVideo(channelId, url, videoRepository, transaction);
 
             await transaction.CommitAsync();
         }
 
         Guid indexTaskId;
+
         await using (var transaction = await connection.OpenAndBeginTransaction())
         {
-            indexTaskId = await repository.AddIndexTask("https://example.org", _libraryId, _userId, transaction);
+            indexTaskId = await repository.AddTask(TaskEntity.Index(_libraryId, _userId, url), transaction);
             await transaction.CommitAsync();
         }
 
@@ -106,7 +74,7 @@ public sealed class TasksTests(ServerFixture fixture) : ServerTests(fixture)
 
         await using (var transaction = await connection.OpenAndBeginTransaction())
         {
-            var taskId = await repository.AddDownloadTask(videoId, _libraryId, _userId, transaction);
+            var taskId = await repository.AddTask(TaskEntity.Download(_libraryId, _userId, videoId), transaction);
 
             var task1 = (await repository.TryDequeueTask(taskId, transaction))!;
             task1.Should().NotBeNull();
@@ -150,19 +118,19 @@ public sealed class TasksTests(ServerFixture fixture) : ServerTests(fixture)
     }
 
     [Test]
-    [Order(2)]
     public async Task CompleteStuckTasks()
     {
         await using var scope = Fixture.Services.CreateAsyncScope();
         var connection = scope.ServiceProvider.GetRequiredService<NpgsqlConnection>();
         var repository = scope.ServiceProvider.GetRequiredService<TaskRepository>();
 
+        var url = $"https://example.org/{Guid.NewGuid():N}";
         Guid taskId;
         Guid taskRunId;
 
         await using (var transaction = await connection.OpenAndBeginTransaction())
         {
-            taskId = await repository.AddIndexTask("https://example.org", _libraryId, _userId, transaction);
+            taskId = await repository.AddTask(TaskEntity.Index(_libraryId, _userId, url), transaction);
             taskRunId = await repository.AddTaskRun(taskId, TaskSource.User, transaction);
             await repository.StartTaskRun(taskRunId, CancellationToken.None);
 
@@ -192,5 +160,120 @@ public sealed class TasksTests(ServerFixture fixture) : ServerTests(fixture)
             runningTask.RunState.Should().Be(RunState.Finished);
             runningTask.Result.Should().Be(TaskResult.Failed);
         }
+    }
+
+    [Test]
+    public async Task DuplicateTasks()
+    {
+        await using var scope = Fixture.Services.CreateAsyncScope();
+        var connection = scope.ServiceProvider.GetRequiredService<NpgsqlConnection>();
+        var repository = scope.ServiceProvider.GetRequiredService<TaskRepository>();
+
+        var url1 = $"https://example.org/{Guid.NewGuid():N}";
+        var url2 = $"https://example.org/{Guid.NewGuid():N}";
+        Guid taskRunId;
+
+        await using (var transaction = await connection.OpenAndBeginTransaction())
+        {
+            var taskId = await repository.AddTask(TaskEntity.Index(_libraryId, _userId, url1), transaction);
+            taskRunId = await repository.AddTaskRun(taskId, TaskSource.User, transaction);
+
+            var otherTaskId = await repository.AddTask(TaskEntity.Index(_libraryId, _userId, url2), transaction);
+            var otherTaskRunId = await repository.AddTaskRun(otherTaskId, TaskSource.User, transaction);
+            await repository.StartTaskRun(otherTaskRunId, CancellationToken.None);
+
+            await transaction.CommitAsync();
+        }
+
+        await using (var transaction = await connection.OpenAndBeginTransaction())
+        {
+            var secondTaskId = await repository.TryAddTask(TaskEntity.Index(_libraryId, _userId, url1), transaction);
+            secondTaskId.Should().BeNull();
+            await transaction.CommitAsync();
+        }
+
+        await repository.StartTaskRun(taskRunId, CancellationToken.None);
+
+        await using (var transaction = await connection.OpenAndBeginTransaction())
+        {
+            var secondTaskId = await repository.TryAddTask(TaskEntity.Index(_libraryId, _userId, url1), transaction);
+            secondTaskId.Should().BeNull();
+            await transaction.CommitAsync();
+        }
+
+        await repository.CompleteTask(taskRunId, CancellationToken.None);
+
+        var ids = await Task.WhenAll(AddTask(url1), AddTask(url1), AddTask(url1));
+        ids.Should().ContainSingle(id => id.HasValue);
+    }
+
+    private async Task<Guid?> AddTask(string url)
+    {
+        await using var scope = Fixture.Services.CreateAsyncScope();
+        var connection = scope.ServiceProvider.GetRequiredService<NpgsqlConnection>();
+        var repository = scope.ServiceProvider.GetRequiredService<TaskRepository>();
+
+        return await connection.ExecuteWithinTransaction(
+            NullLogger.Instance,
+            async transaction =>
+            {
+                var taskId = await repository.TryAddTask(TaskEntity.Index(_libraryId, _userId, url), transaction);
+                if (taskId.HasValue)
+                {
+                    await repository.AddTaskRun(taskId.Value, TaskSource.User, transaction);
+                }
+
+                return taskId;
+            });
+    }
+
+    private async Task<Guid> CreateChannel(ChannelRepository repository, NpgsqlTransaction transaction)
+    {
+        var id = await repository.AddAsync(
+            new()
+            {
+                CreatedByUserId = _userId,
+                ModifiedByUserId = _userId,
+                OwnerId = _userId,
+                Name = Guid.NewGuid().ToString("N"),
+                StoragePath = string.Empty,
+                ExternalId = string.Empty,
+                ExternalUrl = string.Empty,
+                Availability = ExternalAvailability.Public,
+            },
+            transaction);
+
+        return id.Should().HaveValue().And.Subject!.Value;
+    }
+
+    private async Task<Guid> CreateVideo(
+        Guid channelId,
+        string url,
+        VideoRepository repository,
+        NpgsqlTransaction transaction)
+    {
+        var id = await repository.AddAsync(
+            new()
+            {
+                CreatedByUserId = _userId,
+                ModifiedByUserId = _userId,
+                OwnerId = _userId,
+                Name = Guid.NewGuid().ToString("N"),
+                Description = string.Empty,
+                Categories = [],
+                Tags = [],
+                Type = VideoType.Video,
+                ChannelId = channelId,
+                StoragePath = string.Empty,
+                ExternalId = string.Empty,
+                ExternalUrl = url,
+                PublishedAt = SystemClock.Instance.GetCurrentInstant(),
+                RefreshedAt = SystemClock.Instance.GetCurrentInstant(),
+                Availability = ExternalAvailability.Public,
+                Duration = Period.FromMinutes(10),
+            },
+            transaction);
+
+        return id.Should().HaveValue().And.Subject!.Value;
     }
 }